Подготовка к деплою: удалена страница project-diagram (модуль, роут, неактуальные доки), починен отсутствующий импорт useEffect в TasksView, изолирован exceljs от лишней загрузки в ws-task-report
