--- app/(dashboard)/project-diagram/page.tsx
@@ -1,7 +1,0 @@
-import { ProjectDiagram } from '@/modules/project-diagram'
-
-export const dynamic = 'force-dynamic'
-
-export default function ProjectDiagramPage() {
-  return <ProjectDiagram />
-}
