--- modules/project-diagram/index.ts
@@ -1,1 +1,0 @@
-export { ProjectDiagram } from './components/ProjectDiagram'
